+++ src/app/dashboard/details/patients/patient-list/patient-list.component.ts
@@ -3,6 +3,8 @@ import { Patient } from '../../../../model/patient.model';
 import { ModalService } from '../../../../services/modal.service';
 import { PatientService } from '../../../../services/patient.service';
 import { PatientEditComponent } from '../patient-edit/patient-edit.component';
+import { ConfirmContentComponent } from '../../../../shared/confirm-content/confirm-content.component';
+import { ConfirmService } from '../../../../shared/confirm-content/confirm.service';
 
 @Component({
   selector: 'app-patient-list',
@@ -12,18 +14,18 @@ import { PatientEditComponent } from '../patient-edit/patient-edit.component';
 export class PatientListComponent implements OnInit {
   @Input() hospitalId;
   patients:Patient[];
-  constructor(private patientService:PatientService,private modalService:ModalService) { }
+  constructor(private patientService:PatientService,private modalService:ModalService,private confirmService:ConfirmService) { }
 
   ngOnInit() {
 
     this.patientService.patientsChanged.subscribe(
       (patients: Patient[]) => {
-        debugger
+        
         this.patients = patients;
 
     });
     this.patientService.getPatientsByHospital(this.hospitalId).subscribe((result)=>{
-        if(result.total!=0)
+      if(result && result.total!=0)
       {
         this.patients=result.data;
       }
@@ -36,19 +38,24 @@ export class PatientListComponent implements OnInit {
 
   onDeletePatient(patient:Patient)
   {
-
-    this.patientService.deletePatient(patient.id).subscribe(result=>{
-    if(result.success)
-    {
-      this.patientService.getPatientsByHospital(this.hospitalId).subscribe(result=>{
-        if(result.total!=0)
-        {
-          let patients:Patient[]=result.data;
-          this.patientService.patientsChanged.next(patients);
-        }
-      });
-    }
-    });
+    this.confirmService.open(ConfirmContentComponent).then((result) => {
+      let action = `${result}`;
+       if(action.indexOf('yes')!=-1)
+       {
+          this.patientService.deletePatient(patient.id).subscribe(result=>{
+          if(result.success)
+          {
+            this.patientService.getPatientsByHospital(this.hospitalId).subscribe(result=>{
+              if(result && result.total!=0)
+              {
+                let patients:Patient[]=result.data;
+                this.patientService.patientsChanged.next(patients);
+              }
+            });
+          }
+          });
+      }
+      });  
   } 
 }
